samples/python-ml-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
requests==2.31.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2